requirements.txt
crewai
crewai-tools
crewai[tools]
python-dotenv
openai 

# conda create -n fin_analyst_multi_agent python=3.13
# conda activate fin_analyst_multi_agent

# crewai create crew financial_analyst

#remove conda env
# conda remove --name fin_analyst_multi_agent --all